add DBManagerTestFixture and update delete_habit add docstrings
