+++ src/habit_server/db_manager.py
@@ -5,6 +5,7 @@ from sqlalchemy import and_
 from habit_server.db_models import User, UserActivity, UserHabit
 from habit_server.utils import is_valid_email_addr
 
+
 class DBManager():
     """Database Manager for habit server."""
 
@@ -14,6 +15,7 @@ class DBManager():
 
     def add_user(self, user):
         """Add a new user to the database.
+
         :param user User: user to add
         :returns Result: operation result, Ok or Err.
         """
@@ -39,6 +41,7 @@ class DBManager():
 
     def does_user_exist(self, username):
         """Check that a username exist in the database.
+
         :param username str: username to check.
         :returns bool: does user exist in database?
         """
@@ -51,6 +54,7 @@ class DBManager():
 
     def add_habit(self, habit):
         """Add a habit to the databse for a particular user.
+
         :param habit UserHabit: habit to add
         :returns Result: operation result, Ok or Err
         """
@@ -67,6 +71,7 @@ class DBManager():
 
     def get_habits(self, user):
         """Get habits for a particular user.
+
         :param user User: [description]
         :returns Result: operation result, Ok or Err
         """
@@ -82,6 +87,7 @@ class DBManager():
 
     def delete_habit(self, habit):
         """Delete habit for a particular user.
+
         :param habit UserHabit: user habit to delete.
         :returns Result: operation result, Ok or Err
         """
@@ -94,17 +100,14 @@ class DBManager():
         if len(habits) == 0:
             return Result.Err("Habit does not exist, cannot delete")
         else:
-            habit = self._session.query(UserHabit).filter_by(
-            username=habit.username,
-            habitname=habit.habitname
-            ).first()
-            self._session.delete(habit)
+            self._session.delete(habits[0])
             self._session.commit()
 
             return Result.Ok()
 
     def does_habit_exist(self, habit):
         """Check whether a given habit exists for a given user.
+
         :param habit UserHabit: user habit to check existence for
         :returns bool: does the habit exist in the database?
         """
@@ -120,6 +123,7 @@ class DBManager():
 
     def add_activity(self, activity):
         """Add a new activity log for a given user and habit.
+
         :param activity UserActivity: activity to add
         :return Result: operation result, Ok or Err
         """
@@ -139,6 +143,7 @@ class DBManager():
 
     def get_activities(self, habit, trailing_days=100):
         """Get all the activities for a particular habit.
+
         :param habit UserHabit: grab all activities linked to this habit.
         :param trailing_days Optional[int]: filter the activities to last
             trailing_days number of days.
