+++ test/test_api.py
@@ -4,31 +4,9 @@ import pytest
 from flask import Flask, session, json
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import create_engine
-from habit_server.db_manager import DBManager
+from habit_server.app import app, db
 from habit_server.db_models import User, UserActivity, UserHabit
-from habit_server.__init__ import app, db
-
-class DBManagerTestFixture():
-    """Database manager test fixture."""
-    def __enter__(self):
-        """Initialize test app, db manager, db session.
-
-        :return DMManager: initialized database manager.
-        """
-        app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:'
-        with app.app_context():
-            db.create_all()
-            return DBManager(db.session)
-
-    def __exit__(self, exception_type, exception_value, traceback):
-        """ Clean up database session.
-
-        :param exception_type: unused
-        :param exception_value: unused
-        :param traceback: unused
-        """
-        db.session.remove()
-        db.drop_all()
+from DBManagerTestFixture import DBManagerTestFixture
 
 def test_user():
     with DBManagerTestFixture() as db_man:
